Add a component to the Home Page :
Normally, here, we could have imported the MenuList component from where at should have been declared : menu-list.module.ts
But in this version of Ionic, all '.module.ts' has been removed for being handled by 'components.module.ts'. It's there that the components are exported, it's a factory.
So when we need to import a component, we should import the ComonentsModule under the path where components.module.ts is